moved the request to the vmd to a data accessor file
